Bug Fixes and BallGame

+++ PragmaEngine/IOManager.cpp
@@ -12,11 +12,11 @@ namespace PragmaEngine {
 		file.seekg(0, std::ios::end);
 
 		//Get the file size
-		int fileSize = file.tellg();
+		unsigned int fileSize = (unsigned int)file.tellg();
 		file.seekg(0, std::ios::beg);
 
 		//Reduce the file size by any header bytes that might be present
-		fileSize -= file.tellg();
+		fileSize -= (unsigned int)file.tellg();
 
 		buffer.resize(fileSize);
 		file.read((char *)&(buffer[0]), fileSize);
